Issue #5022: Use react-gravatar plugin
And move Comment component into its own file
