+++ _comments-app/components/Comments.js
@@ -1,8 +1,8 @@
 import React, {Component} from 'react';
 import api from '../utils/api';
 import PropTypes from 'prop-types';
+import Comment from './Comment';
 import Message from './Message';
-// import Gravatar from './Gravatar';
 
 function CommentFormLink (props) {
   'use strict';
@@ -21,60 +21,6 @@ function CommentFormLink (props) {
 
 CommentFormLink.propTypes = {
   onClick: PropTypes.func.isRequired
-};
-
-class Comment extends Component {
-  /**
-   * Convert encoded HTML before displaying a comment.
-   *
-   * @param commentText
-   */
-  static decodeHtml(commentText) {
-    const parser = new DOMParser;
-    const dom = parser.parseFromString(
-      '<!doctype html><body>' + commentText,
-      'text/html');
-    return dom.body.textContent;
-  }
-
-  render () {
-    const commentText = Comment.decodeHtml(this.props.comment);
-    return (
-      // <li className={this.props.class}>
-      //   {this.props.savasian === '1' && <img src='/assets/img/logo.svg' className='comment__logo' alt='Savas Labs logo' />}
-      //   <p className="comment__name"><span className="c-magenta">{this.props.name}</span> says:</p>
-      //   <p className="comment__date">{this.props.date}</p>
-      //   <p className="comment__text">{commentText}</p>
-      // </li>
-      <li className={this.props.class}>
-        <div className="comment__avatar">
-          <div className="comment__avatar--image">{this.props.savasian === '1' && <img src='/assets/img/logo.svg' className='comment__logo' alt='Savas Labs logo' />}</div>
-          <div className="comment__avatar--bg"></div>
-        </div>
-        <div className="comment__content">
-          <div className="comment__content--header">
-            <div className="comment__content--header__name">{this.props.name}</div>
-            <div className="comment__content--header__actions">
-              <a href="#reply" alt="Reply" aria-hidden="true"></a>
-              <a href="#" alt="Copy Permalink" aria-hidden="true"></a>
-            </div>
-          </div>
-          <div className="comment__content--body">
-            <div className="comment__content--body__text">{commentText}</div>
-            <div className="comment__content--body__date">{this.props.date}</div>
-          </div>
-        </div>
-      </li>
-    );
-  }
-}
-
-Comment.propTypes = {
-  class: PropTypes.string.isRequired,
-  savasian: PropTypes.string.isRequired,
-  name: PropTypes.string.isRequired,
-  date: PropTypes.string.isRequired,
-  comment: PropTypes.string.isRequired
 };
 
 class Comments extends Component {
@@ -143,6 +89,7 @@ class Comments extends Component {
             if (comment.savasian === 1) {
               commentClass = 'comment savasian';
             }
+
             return (
               <Comment
                 key={index}
@@ -151,6 +98,7 @@ class Comments extends Component {
                 name={comment.name}
                 date={comment.created_at}
                 comment={comment.comment}
+                emailHash={comment.hash}
               />
             );
           })}
