+++ package.json
@@ -28,19 +28,19 @@
     "babel-preset-env": "^1.6.0",
     "babel-preset-es2015": "^6.24.1",
     "babel-preset-react": "^6.24.1",
-    "browser-sync": "^2.14.0",
+    "browser-sync": "^2.23.6",
     "del": "^2.2.2",
-    "eslint": "^4.9.0",
-    "eslint-plugin-react": "^7.4.0",
+    "eslint": "^4.16.0",
+    "eslint-plugin-react": "^7.6.1",
     "gulp": "^3.9.1",
     "gulp-accessibility": "^3.0.1",
     "gulp-append-prepend": "^1.0.3",
-    "gulp-babel": "^7.0.0",
+    "gulp-babel": "^7.0.1",
     "gulp-cache": "^0.4.6",
     "gulp-clean-css": "^2.0.12",
     "gulp-concat": "^2.6.0",
     "gulp-imagemin": "^3.0.3",
-    "gulp-jshint": "^2.0.1",
+    "gulp-jshint": "^2.1.0",
     "gulp-notify": "^2.2.0",
     "gulp-postcss": "^6.1.1",
     "gulp-rename": "^1.2.2",
@@ -53,7 +53,6 @@
     "imagemin-jpeg-recompress": "^5.1.0",
     "is-retina": "^1.0.3",
     "jshint": "^2.9.3",
-    "md5": "^2.2.1",
     "prop-types": "^15.5.10",
     "qs": "^6.5.1",
     "query-string": "^5.0.0",
@@ -61,6 +60,6 @@
     "react-dom": "^15.6.1",
     "react-gravatar": "^2.6.3",
     "run-sequence": "^1.2.2",
-    "webpack": "^3.5.5"
+    "webpack": "^3.10.0"
   }
 }
